Tidy interpreter
