+++ src/Language/Mimsa/Interpreter/BuiltIns.hs
@@ -1,0 +1,23 @@
+{-# LANGUAGE OverloadedStrings #-}
+
+module Language.Mimsa.Interpreter.BuiltIns where
+
+import Data.Text (Text)
+import qualified Data.Text as T
+import Language.Mimsa.Types
+
+wrappedName :: Variable -> Variable
+wrappedName = appendToVar "__unwrapped"
+
+appendToVar :: Text -> Variable -> Variable
+appendToVar tx (BuiltIn (Name n)) = BuiltIn (Name (n <> tx))
+appendToVar _ a = a
+
+wrappedVarName :: Variable -> Int -> Variable
+wrappedVarName name i = appendToVar label name
+  where
+    label = "__unwrapped" <> T.pack (show i)
+
+unwrap :: Variable -> Maybe Variable
+unwrap (BuiltIn (Name n)) = BuiltIn . Name <$> T.stripSuffix "__unwrapped" n
+unwrap a = Just a
