+++ test/Test/Repl.hs
@@ -212,3 +212,9 @@ spec =
       it "let combine = \\b -> \\a -> \"horse\" in reduceList(combine)(\"\")([1,2,3])" $ do
         result <- eval stdLib "let combine = \\b -> \\a -> \"horse\" in reduceList(combine)(\"\")([1,2,3])"
         result `shouldBe` Right (MTString, str' "horse")
+      it "reduceList(\\aa -> \\bb -> addInt(aa)(bb))(0)([1,2,3,4,5])" $ do
+        result <- eval stdLib "reduceList(\\aa -> \\bb -> addInt(aa)(bb))(0)([1,2,3,4,5])"
+        result `shouldBe` Right (MTInt, int 14)
+      it "addInt(1)(addInt(addInt(2)(4))(5))" $ do
+        result <- eval stdLib "addInt(1)(addInt(addInt(2)(4))(5))"
+        result `shouldBe` Right (MTInt, int 12)
